Move column path arg retrieval to a function
Also improves errors
